refactor fabric handles
